+++ src/frontend/components/ComponentSheet.tsx
@@ -1,8 +1,10 @@
-'use client';
+'use client'
 
-import { Button } from "./ui/button";
-import ComponentViewer from "./ComponentViewer";
-import { ComponentData } from "./models";
+import { useState } from "react"
+import { Button } from "./ui/button"
+import ComponentViewer from "./ComponentViewer"
+import ComponentViewerSkeleton from "./ComponentViewerSkeleton"
+import { ComponentData } from "./models"
 import { 
   Sheet, 
   SheetClose,
@@ -11,10 +13,36 @@ import {
   SheetFooter,
   SheetHeader,
   SheetTitle,
-  SheetTrigger } from "./ui/sheet";
-import { Card, CardContent, CardHeader, CardTitle } from "./ui/card";
-import { rgbToHex } from "@/lib/utils";
-import { Tooltip, TooltipProvider, TooltipTrigger, TooltipContent } from "./ui/tooltip";
+  SheetTrigger } from "./ui/sheet"
+import { Card, CardContent, CardHeader, CardTitle } from "./ui/card"
+import { rgbToHex } from "@/lib/utils"
+import { Tooltip, TooltipProvider, TooltipTrigger, TooltipContent } from "./ui/tooltip"
+import { hexComponentColor, componentColorString } from "@/lib/utils"
+
+interface FetchComponentGeometryProps {
+  geometry_id: string
+}
+
+const fetch_component_geometry = async ({ geometry_id }: FetchComponentGeometryProps) => {
+  const params = new URLSearchParams(
+    { geometry_id }
+  )
+  const response = await fetch(
+    `${process.env.NEXT_PUBLIC_BASE_URL}/api/fetch-component-geometry?${params.toString()}`,
+    {
+      method: 'GET',
+      cache: 'no-cache'
+    }
+  )
+  if (!response.ok) {
+    throw new Error('Failed to fetch component geometry')
+  }
+  return response.json()
+}
+
+function inject_geometry(component: ComponentData, component_geometry: any) {
+      component.geometry = component_geometry
+}
 
 export default function ComponentSheet({
   component_data,
@@ -23,20 +51,37 @@ export default function ComponentSheet({
 }) {
 
   // Component Color
-  const colR = Math.round(component_data.color[0]);
-  const colG = Math.round(component_data.color[1]);
-  const colB = Math.round(component_data.color[2]);
-  const hexcol = rgbToHex(colR, colG, colB)
+  const component_color_str = componentColorString(component_data.color)
+  const component_color_hex = hexComponentColor(component_data.color)
+
+  // Geometry Fetching
+  const [fetchedData, setFetchedData] = useState(null);
+  const [isLoading, setIsLoading] = useState(false);
+  const [isComponentViewerVisible, setIsComponentViewerVisible] = useState(false);
+
+  const handleButtonClickComponentDetail = async () => {
+    setIsLoading(true);
+    setIsComponentViewerVisible(false); // Hide ComponentViewer while fetching
+    try {
+      const component_geometry = await fetch_component_geometry({ geometry_id: component_data._id });
+      inject_geometry(component_data, component_geometry.geometry);
+      setFetchedData(component_geometry);
+      setIsComponentViewerVisible(true); // Show ComponentViewer once data is fetched
+    } catch (error) {
+      console.error("Error fetching data:", error);
+    } finally {
+      setIsLoading(false);
+    }
+  }
 
   return (
-    
     <Sheet>
       <SheetTrigger asChild>
         <div className='text-left align-text-top cursor-pointer'>
           <TooltipProvider>
             <Tooltip>
               <TooltipTrigger asChild>
-                <Button variant="ghost" className="h-5 w-[315px] hover:bg-[#009cda] hover:text-white">{component_data._id}</Button>
+                <Button variant="ghost" className="h-5 w-[315px] hover:bg-[#009cda] hover:text-white" onClick={handleButtonClickComponentDetail}>{component_data._id}</Button>
               </TooltipTrigger>
               <TooltipContent>
                 Click to preview
@@ -60,14 +105,15 @@ export default function ComponentSheet({
                 Material Thickness: {component_data.materialthickness} <br/>
                 <div className="flex items-center max-w-12">
                 Color: 
-                <div className="ml-2 avatar rounded-full min-h-4 min-w-4 max-w-5 max-h-5 items-center justify-left" style={{backgroundColor: hexcol}}></div>
-                <div className="ml-2 items-center justify-center text-center">{colR}/{colG}/{colB}</div>
+                <div className="ml-2 avatar rounded-full min-h-4 min-w-4 max-w-5 max-h-5 items-center justify-left" style={{backgroundColor: component_color_hex}}></div>
+                <div className="ml-2 items-center justify-center text-center">{component_color_str}</div>
                 </div>
               </CardContent>
             </Card>
           </SheetDescription>
         </SheetHeader>
-            <ComponentViewer component_data={component_data}/>
+        {isLoading ? <ComponentViewerSkeleton />: 
+        isComponentViewerVisible && <ComponentViewer component_data={component_data} />}
         <SheetFooter>
           <div >
           <SheetClose asChild className="text-center items-center">