+++ src/frontend/components/Sidebar.tsx
@@ -6,7 +6,7 @@ import AppMenu from "./AppMenu";
 export default function Sidebar() {
 
   return (
-    <div className="flex flex-col gap-4 w-[300px] min-w-[300px] p-4 h-full justify-between">
+    <div className="fixed flex flex-col gap-4 w-[300px] min-w-[300px] p-4 h-full justify-between">
       
       {/* <div>
         <UserItem/>