+++ CHANGELOG.md
@@ -69,3 +69,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Credits Page `app/credits/page.tsx`
 - Added Footer Component `components/Footer.tsx`
 - Added Function `copyright_date` to `lib/utils.ts`
+
+## [0.0.1.4] - 2024-06-04
+
+### Versions
+
+- CSC FastAPI Backend:  0.1.0.2
+- CSC React Frontend:   0.1.4.2
+- CSC Sheetscan Module: 0.0.1.10
+
+### Added
+
+#### CSC React Frontend
+
+- Added lazy loading of components
+- Added `fetch-components-shallow` route to API
+- Added `fetch-component-geometry` route to API
+
+#### CSC FastAPI Backend:
+
+- Added API endpoints for retrieving component geometry and shallow components without geometry
+
+### Changed
+
+#### CSC React Frontend
+
+- Components in the overview are now retrieved _shallow_, e.g. without geometry.
+- Geometry is loaded on click during opening of the Component Detail Sheet.